OK. When I go here: http://ucjeps.berkeley.edu/cgi-bin/get_consort.pl?taxon_name=Boechera%20elkoensis and click on BerkeleyMapper with layers (or without) and on the ensuing map, under MarkerClusterer, select placemarks, the cluster symbol changes to 3 navy-blue markers, when what I expect is one royal blue, one navy, and one light gray
I get the same results on Mac 10.6.8 using Safari 5.1.8 or Chrome
26.0.1410.63
and on 10.7.something using Safari
and on 10.5.8 using Firefox


John Deck <jdeck@berkeley.edu>
4:40 PM (0 minutes ago)

to Richard 
what fun.  google maps api release broke this.  Not so much their fault since the styled marker library i use to apply the colors is a 3rd party app which google doesn't support.  here is a summary of releases and what works/doesn't

3.9 works
3.10 doesn't work (stable release)
3.11 doesn't work (current release)
3.12 works (experimental release).

3.9 will be end of life pretty soon so i went with 3.12/"experimental" since i assume it will shortly be stable and the rest of the features i need seems to work